9th

+++ models/seed.js
@@ -12,19 +12,19 @@ const db = mongoose.connection
 db.on('open', () => {
     // Bringing in the array for fishes
     const startFishs = [
-        {name: "Tarpon", BodyOfWater: "Freshwater", readyToCatch: false},
-        {name: "Snook", BodyOfWater: "Euryhaline", readyToCatch: false},
-        {name: "Redfish", BodyOfWater: "Saltwater", readyToCatch: false},
-        {name: "Speckled Trout", BodyOfWater: "Salt-water", readyToCatch: false},
-        {name: "Florida Largemouth Bass", BodyOfWater: "Freshwater", readyToCatch: false},
-        {name: "Sheepshead", BodyOfWater: "Salt-water", readyToCatch: false},
-        {name: "Kingfish", BodyOfWater: "Salt-water", readyToCatch: false},
-        {name: "Gulf Flounder", BodyOfWater: "Salt-water", readyToCatch: false},
-        {name: "Goliath Grouper", BodyOfWater: "Salt-water", readyToCatch: false},
-        {name: "Bonnethead Shark", BodyOfWater: "Salt-water", readyToCatch: false},
+        {name: "Tarpon", bodyOfWater: "Freshwater", readyToCatch: false},
+        {name: "Snook", bodyOfWater: "Euryhaline", readyToCatch: false},
+        {name: "Redfish", bodyOfWater: "Saltwater", readyToCatch: false},
+        {name: "Speckled Trout", bodyOfWater: "Salt-water", readyToCatch: false},
+        {name: "Florida Largemouth Bass", bodyOfWater: "Freshwater", readyToCatch: false},
+        {name: "Sheepshead", bodyOfWater: "Salt-water", readyToCatch: false},
+        {name: "Kingfish", bodyOfWater: "Salt-water", readyToCatch: false},
+        {name: "Gulf Flounder", bodyOfWater: "Salt-water", readyToCatch: false},
+        {name: "Goliath Grouper", bodyOfWater: "Salt-water", readyToCatch: false},
+        {name: "Bonnethead Shark", bodyOfWater: "Salt-water", readyToCatch: false}
     ]
 
-     Fish.deleteMany({})
+     Fish.deleteMany({owner: null})
          .then(deletedFishs => {
              console.log('this is what .delteMany returns', deletedFishs)
 
@@ -34,6 +34,7 @@ db.on('open', () => {
                 db.close()
             })
             .catch(error => {
+                console.log('error inside fish create promise', error)
                 db.close()
             })
         })
